backend/requirements.txt
# Student LMS Backend Requirements
# Production-ready with Authentication

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
supabase==2.3.0
httpx==0.26.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Utilities
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
